html/semantics/embedded-content/the-img-element/nonexistent-image.html

<!DOCTYPE HTML>
<meta charset="utf-8">
<title>Loading a nonexistent image should fail</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<img id="img1">
<img id="img2">
<img id="img3">

<script>
  async_test(t => {
    var img = document.getElementById("img1");
    img.onload = t.unreached_func("image.onload() was not supposed to be called");
    img.onerror = t.step_func_done(e => {
      assert_equals(e.type, "error", "image.onerror() called");
    });
    img.src = "404";
  }, "Loading a no suffix and nonexistent image should fire onerror events");

  async_test(t => {
    var img = document.getElementById("img2");
    img.onload = t.unreached_func("image.onload() was not supposed to be called");
    img.onerror = t.step_func_done(e => {
      assert_equals(e.type, "error", "image.onerror() called");
    });
    img.src = "404.png";
  }, "Loading a nonexistent image should fire onerror events");

  async_test(t => {
    var img = document.getElementById("img3");
    img.onload = t.unreached_func("image.onload() was not supposed to be called");
    img.onerror = t.step_func_done(e => {
      assert_equals(e.type, "error", "image.onerror() called");
    });
    img.src = "";
  }, "Loading an empty URL image should fire onerror events");
</script>
